added some reducers/actions
